Moved the processing of the request for showing the admins' list from 'index.js' to 'chat.js'
